More optimizations to the base_rules makefile.
This saves another 5-15% of makefile eval time (depending on
target/branch).

I'm not sure if deps-license is still used at all, but we'll only save
the information for it if it's actually going to run now.

Since the default module tag is "optional", we really don't need to save
a list of most modules in another variable. I don't see anything that
really needs to see the list of optional modules.

Test: treehugger
Change-Id: Iaf54e9932fbe8e61f10bd0c9336fc4fdefea27ab